Split join operator suite
